backend/requirements.txt
asgiref==3.11.1
python-dotenv==1.0.1
black==24.4.2
click==8.3.1
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==7.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.8.0
psycopg2-binary==2.9.9
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.11.0
pytokens==0.4.1
pytz==2025.2
sqlparse==0.5.5
gunicorn
reportlab==4.2.5
openpyxl==3.1.5
requests>=2.31.0
coverage>=7.0
redis>=5.0
redis>=5.0